requirements.txt
# BDA-AAT Dependencies
tweepy>=4.14.0
yfinance>=0.2.36
pymongo>=4.6.1
pandas>=2.1.4
numpy>=1.26.2
scikit-learn>=1.3.2
transformers>=4.36.2
torch>=2.2.0
plotly>=5.18.0
matplotlib>=3.8.2
seaborn>=0.13.0
streamlit>=1.29.0
python-dotenv>=1.0.0
tqdm>=4.66.1
